requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
google-genai>=0.3.0
PyMuPDF>=1.23.0
pandas>=2.0.0
